% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon_table.R
\name{ccdB_wt_nucleotide}
\alias{ccdB_wt_nucleotide}
\title{CcdB Wild-Type Nucleotide Sequence}
\usage{
ccdB_wt_nucleotide()
}
\value{
Character string (306 nt).
}
\description{
Returns the CcdB wild-type nucleotide sequence used as the
reference in the original cysteine scanning library.
}
