% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon_table.R
\name{ccdB_wt_protein}
\alias{ccdB_wt_protein}
\title{CcdB Wild-Type Protein Sequence}
\usage{
ccdB_wt_protein()
}
\value{
Character string (101 aa + stop).
}
\description{
Returns the translated CcdB wild-type protein sequence.
}
