% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta_convert.R
\name{reverse_complement_seq}
\alias{reverse_complement_seq}
\title{Reverse Complement a DNA Sequence}
\usage{
reverse_complement_seq(seq)
}
\arguments{
\item{seq}{Character string of DNA.}
}
\value{
Character string of the reverse complement.
}
\description{
Computes the reverse complement of a DNA sequence string.
}
\examples{
reverse_complement_seq("ATGC")  # "GCAT"
}
