% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon_table.R
\name{translate_sequence}
\alias{translate_sequence}
\title{Translate Nucleotide Sequence to Protein}
\usage{
translate_sequence(nt_seq)
}
\arguments{
\item{nt_seq}{Character string of the nucleotide sequence (length must be
divisible by 3).}
}
\value{
Character string of the amino acid sequence.
}
\description{
Translates a nucleotide sequence to amino acid sequence using
the standard genetic code.
}
\examples{
translate_sequence("ATGCAGTTTAAG")  # "MQFK"
}
