% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta_convert.R
\name{write_fasta}
\alias{write_fasta}
\title{Write FASTA Data Frame to File}
\usage{
write_fasta(fasta_df, file)
}
\arguments{
\item{fasta_df}{Data frame with \code{fasta_header} and
\code{fasta_sequence} columns.}

\item{file}{Path to the output file.}
}
\value{
Invisible NULL. Called for side effect.
}
\description{
Writes a FASTA data frame to a file.
}
